use usize

+++ src/output.rs
@@ -22,18 +22,18 @@ mod ioctl {
     const  DISP_DEVICE_SWITCH: u32       = 0x0F;
     const  DISP_GET_OUTPUT_TYPE: u32     = 0x09;
 
-    ioctl_write_ptr_bad!(get_output, DISP_GET_OUTPUT_TYPE, [u32]); 
-    ioctl_write_ptr_bad!(set_output, DISP_DEVICE_SWITCH, [u32]);   
+    ioctl_write_ptr_bad!(get_output, DISP_GET_OUTPUT_TYPE, [usize]); 
+    ioctl_write_ptr_bad!(set_output, DISP_DEVICE_SWITCH, [usize]);   
 }
 
 pub fn get_output(fd: RawFd) -> io::Result<OutputType> {
-    let data: [u32; 4] = [0, 0, 0, 0];
+    let data: [usize; 4] = [0, 0, 0, 0];
     let raw_value: u32 = from_nix_result(unsafe { ioctl::get_output(fd, &data) })?.try_into().unwrap();
     Ok(if raw_value == 1 { OutputType::LCD } else { OutputType::HDMI })
 }
 
 pub fn set_output(fd: RawFd, output_type: OutputType) -> io::Result<()> {
-    let data: [u32; 4] = match output_type {
+    let data: [usize; 4] = match output_type {
         OutputType::LCD => {[0, 1, 0, 0]}
         OutputType::HDMI => {[0, 4, 5, 0]}
     };
